+++ crates/cubek-std/tests/std/tile_dsl/space/point.rs
@@ -6,7 +6,7 @@ use cubecl::prelude::*;
 /// need only their own axes.
 #[derive(CubeType)]
 pub struct Point {
-    coords: Sequence<u32>,
+    coords: Sequence<usize>,
     #[cube(comptime)]
     frame: Space,
 }
@@ -14,12 +14,12 @@ pub struct Point {
 #[cube]
 impl Point {
     /// Wrap per-axis runtime coordinates (in `frame` order) as a point.
-    pub fn new(coords: Sequence<u32>, #[comptime] frame: Space) -> Point {
+    pub fn new(coords: Sequence<usize>, #[comptime] frame: Space) -> Point {
         Point { coords, frame }
     }
 
     /// The coordinate along `axis`, located via the point's frame.
-    pub fn get(&self, #[comptime] axis: Axis) -> u32 {
+    pub fn get(&self, #[comptime] axis: Axis) -> usize {
         *self.coords.index(comptime!(self.frame.position(axis)))
     }
 }
